Remove padding override on global container-fluid class.
It's better to be explicit in overriding the padding of container-fluid
with an extra class than to affect the way this global class works.
It is used more frequently in bootstrap 3 in areas not related to the grid,
as in the navbar. Bootstrap 3 also changed the way padding and margins work
together with rows.