Load bootstrap 3 files by default and rename old ones to v2.
The dashboard will still use Bootstrap v2, the responsive.less
references the less files which are now replaced by v3 thus the
reference needs to be removed for now. It is replaced by the
minimized responsive css included in bootstrap 2.
This will make it hard to customize the dashboard until it is
migarted.